perf: Prosemirror deps loaded with Document model
