+++ app/models/Document.ts
@@ -3,9 +3,6 @@ import i18n, { t } from "i18next";
 import capitalize from "lodash/capitalize";
 import floor from "lodash/floor";
 import { action, autorun, computed, observable, set } from "mobx";
-import { Node, Schema } from "prosemirror-model";
-import ExtensionManager from "@shared/editor/lib/ExtensionManager";
-import { richExtensions, withComments } from "@shared/editor/nodes";
 import type {
   JSONObject,
   NavigationNode,
@@ -17,7 +14,6 @@ import {
   NavigationNodeType,
   NotificationEventType,
 } from "@shared/types";
-import { ProsemirrorHelper } from "@shared/utils/ProsemirrorHelper";
 import Storage from "@shared/utils/Storage";
 import { isRTL } from "@shared/utils/rtl";
 import slugify from "@shared/utils/slugify";
@@ -686,47 +682,6 @@ export default class Document extends ArchivableModel implements Searchable {
       []
     );
   }
-
-  /**
-   * Returns the markdown representation of the document derived from the ProseMirror data.
-   *
-   * @returns The markdown representation of the document as a string.
-   */
-  toMarkdown = () => {
-    const extensionManager = new ExtensionManager(withComments(richExtensions));
-    const serializer = extensionManager.serializer();
-    const schema = new Schema({
-      nodes: extensionManager.nodes,
-      marks: extensionManager.marks,
-    });
-
-    const doc = Node.fromJSON(
-      schema,
-      ProsemirrorHelper.attachmentsToAbsoluteUrls(this.data)
-    );
-
-    const markdown = serializer.serialize(doc, {
-      softBreak: true,
-    });
-    return markdown;
-  };
-
-  /**
-   * Returns the plain text representation of the document derived from the ProseMirror data.
-   *
-   * @returns The plain text representation of the document as a string.
-   */
-  toPlainText = () => {
-    const extensionManager = new ExtensionManager(withComments(richExtensions));
-    const schema = new Schema({
-      nodes: extensionManager.nodes,
-      marks: extensionManager.marks,
-    });
-    const text = ProsemirrorHelper.toPlainText(
-      Node.fromJSON(schema, this.data)
-    );
-    return text;
-  };
 
   download = (contentType: ExportContentType) =>
     client.post(
