+++ shared/i18n/locales/en_US/translation.json
@@ -216,7 +216,6 @@
   "Deleted Collection": "Deleted Collection",
   "Untitled": "Untitled",
   "Unpin": "Unpin",
-  "{{ minutes }}m read": "{{ minutes }}m read",
   "Select a location to copy": "Select a location to copy",
   "Document copied": "Document copied",
   "Couldn’t copy the document, try again?": "Couldn’t copy the document, try again?",
@@ -345,6 +344,7 @@
   "Reaction picker": "Reaction picker",
   "Could not load reactions": "Could not load reactions",
   "Reaction": "Reaction",
+  "{{ minutes }}m read": "{{ minutes }}m read",
   "Revision deleted": "Revision deleted",
   "Current version": "Current version",
   "{{userName}} edited": "{{userName}} edited",
